Add UEFI application media types.
The following media types have been added to the IANA registry at
http://www.iana.org/assignments/provisional-standard-media-types
and are expected to be added in the next revision of the standard.  This
enables them to be used for network booting on UEFI machines.

The types are:

application/efi			efi
application/vnd.uefi.img	efi.img uefi.img
application/vnd.uefi.iso	efi.iso uefi.iso